4.8 Отрефакторит модуль со вспомогательными функциями и константами
