+++ src/const.js
@@ -8,4 +8,52 @@ const POINT_EMPTY = {
   type: 'flight'
 };
 
-export { POINT_EMPTY };
+const DESCRIPTIONS = [
+  'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
+  'Cras aliquet varius magna, non porta ligula feugiat eget.',
+  'Fusce tristique felis at fermentum pharetra.',
+  'Aliquam id orci ut lectus varius viverra.',
+  'Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.',
+  'Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.'
+];
+
+const CITIES = [
+  'Moscow',
+  'Amsterdam',
+  'Chamonix',
+  'Geneva',
+  'Buenos Aires',
+  'Paris',
+  'Dublin',
+  'Wellington'
+];
+
+const Duration = {
+  HOUR: 5,
+  DAY: 5,
+  MIN: 59
+};
+
+const TYPES = [
+  'taxi',
+  'bus',
+  'train',
+  'ship',
+  'drive',
+  'flight',
+  'check-in',
+  'sightseeing',
+  'restaurant'
+];
+
+const DUFULT_TYPE = 'flight';
+
+
+export {
+  POINT_EMPTY,
+  DESCRIPTIONS,
+  CITIES,
+  Duration,
+  TYPES,
+  DUFULT_TYPE
+};
