--- src/mock/const.js
@@ -1,37 +1,0 @@
-const DESCRIPTIONS = [
-  'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
-  'Cras aliquet varius magna, non porta ligula feugiat eget.',
-  'Fusce tristique felis at fermentum pharetra.',
-  'Aliquam id orci ut lectus varius viverra.',
-  'Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.',
-  'Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.'
-];
-
-const CITIES = [
-  'Moscow',
-  'Amsterdam',
-  'Chamonix',
-  'Geneva'
-];
-
-const Duration = {
-  HOUR: 5,
-  DAY: 5,
-  MIN: 59
-};
-
-const TYPES = [
-  'taxi',
-  'bus',
-  'train',
-  'ship',
-  'drive',
-  'flight',
-  'check-in',
-  'sightseeing',
-  'restaurant'
-];
-
-const DUFULT_TYPE = 'flight';
-
-export { DESCRIPTIONS, CITIES, Duration, TYPES, DUFULT_TYPE };
