+++ src/mock/destination.js
@@ -1,5 +1,5 @@
-import { getRandomArrayElement, getRandomInteger } from './utils.js';
-import { DESCRIPTIONS, CITIES } from './const.js';
+import { getRandomArrayElement, getRandomInteger } from '../utils/common.js';
+import { DESCRIPTIONS, CITIES } from '../const.js';
 
 function generateDestination() {
   return {
