+++ src/utils/common.js
@@ -1,0 +1,15 @@
+function getRandomArrayElement(items) {
+  return items[Math.floor(Math.random() * items.length)];
+}
+
+function getRandomInteger(a = 0, b = 1) {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+
+  return Math.floor(lower + Math.random() * (upper - lower + 1));
+}
+
+export {
+  getRandomArrayElement,
+  getRandomInteger
+};
