+++ src/utils/point.js
@@ -1,0 +1,86 @@
+import dayjs from 'dayjs';
+import duration from 'dayjs/plugin/duration';
+import relativeTime from 'dayjs/plugin/relativeTime';
+import { Duration } from '../const.js';
+import { getRandomInteger } from '../utils/common.js';
+
+
+dayjs.extend(duration);
+dayjs.extend(relativeTime);
+
+const MSEC_IN_SEC = 1000;
+const SEC_IN_MIN = 60;
+const MIN_IN_HOUR = 60;
+const HOUR_IN_DAY = 24;
+const DAY_FORMAT = 'MMM DD';
+const TIME_FORMAT = 'HH:mm';
+const MSEC_IN_HOUR = MIN_IN_HOUR * SEC_IN_MIN * MSEC_IN_SEC;
+const MSEC_IN_DAY = HOUR_IN_DAY * MSEC_IN_HOUR;
+
+function formatStringToDateTime(data) {
+  return dayjs(data).format('YYYY-MM-DDTHH:mm');
+}
+
+function formatStringToShotrDate(data) {
+  return dayjs(data).format('MMM DD');
+}
+
+function formatStringToTime(data) {
+  return dayjs(data).format('HH:mm');
+}
+
+function getPointDuration(dateFrom, dateTo) {
+  const timeDiff = dayjs(dateTo).diff(dayjs(dateFrom));
+
+  let pointDuration = 0;
+
+  switch (true) {
+    case (timeDiff >= MSEC_IN_DAY):
+      pointDuration = dayjs.duration(timeDiff).format('DD[D] HH[H] mm[M]');
+      break;
+    case (timeDiff >= MSEC_IN_HOUR):
+      pointDuration = dayjs.duration(timeDiff).format('HH[H] mm[M]');
+      break;
+    case (timeDiff < MSEC_IN_HOUR):
+      pointDuration = dayjs.duration(timeDiff).format('mm[M]');
+      break;
+  }
+
+  return pointDuration;
+}
+
+let date = dayjs().subtract(getRandomInteger(0, Duration.DAY), 'day').toDate();
+
+function getDate({ next }) {
+  const minsGap = getRandomInteger(0, Duration.MIN);
+  const hoursGap = getRandomInteger(1, Duration.HOUR);
+  const daysGap = getRandomInteger(1, Duration.DAY);
+
+  if (next) {
+    date = dayjs(date)
+      .add(minsGap, 'minute')
+      .add(hoursGap, 'hour')
+      .add(daysGap, 'day')
+      .toDate();
+  }
+
+  return date;
+}
+
+function toDay(dateTime) {
+  return dateTime ? dayjs(dateTime).format(DAY_FORMAT) : '';
+}
+
+function toTime(dateTime) {
+  return dateTime ? dayjs(dateTime).format(TIME_FORMAT) : '';
+}
+
+export {
+  formatStringToDateTime,
+  formatStringToShotrDate,
+  formatStringToTime,
+  getPointDuration,
+  getDate,
+  toDay,
+  toTime
+};
